requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
groq
python-dotenv
plotly
fpdf2
qrcode
pillow
matplotlib
seaborn
